sim api's completed, translation implemented

+++ routes/sockets.js
@@ -585,7 +585,7 @@ sockets.listen = function (server) {
 
             //************** */ SIM MODULE
             socket.on(Constants.ACK_SIM + device_id, (response) => {
-                // console.log('ack ==============> ', response)
+                console.log('ack ==============> ', response)
                 if (response != undefined) {
                     let uQry = `UPDATE sims SET sync = '1' WHERE device_id = '${response.device_id}' AND iccid = '${response.iccid}'`;
                     sql.query(uQry);
@@ -593,18 +593,23 @@ sockets.listen = function (server) {
             })
 
             socket.on(Constants.RECV_SIM + device_id, (response) => {
-                console.log('ack ===== RECV_SIM =========> ', response);
+                console.log('===== RECV_SIM =========> ', response);
                 sockets.updateSimRecord(device_id, response);
             })
 
 
             let sUnEmitSims = `SELECT * FROM sims WHERE sync = '0'`;
+            // console.log('========= check data when socket => re-connect ================= ', sUnEmitSims);
             let simResult = await sql.query(sUnEmitSims);
+            // console.log('results are: ', simResult);
             if (simResult.length > 0) {
                 simResult.forEach(async function (data, index) {
+                    // data['guest'] = data.guest == 1 ? true : false;
+                    // data['encrypt'] = data.encrypt == 1 ? true : false;
+                    // console.log('updated result is: ', data);
                     socket.emit(Constants.SEND_SIM + data.device_id, {
                         device_id: data.device_id,
-                        sim: (data === undefined || data === null || data === '') ? '{}' : data,
+                        sim: (data === undefined || data === null || data === '') ? '{}' : JSON.stringify(data),
                     });
                     let uQry = `UPDATE sims SET sync = '1' WHERE device_id = '${data.device_id}' AND iccid = '${data.iccid}'`;
                     await sql.query(uQry);
@@ -676,11 +681,17 @@ sockets.listen = function (server) {
 }
 
 
-sockets.sendRegSim = async (data) => {
-    console.log('sendRegSim data is: ', data)
-    io.emit(Constants.SEND_SIM + data.device_id, {
-        device_id: data.device_id,
-        sim: (data === undefined || data === null || data === '') ? '{}' : data,
+sockets.sendRegSim = async (device_id, action, data) => {
+    console.log('sendRegSim data is=> ', {
+        action,
+        device_id,
+        entries: (data === undefined || data === null || data === '') ? '{}' : JSON.stringify(data),
+    });
+
+    io.emit(Constants.SEND_SIM + device_id, {
+        action,
+        device_id,
+        entries: (data === undefined || data === null || data === '') ? '{}' : JSON.stringify(data),
     });
 }
 
@@ -689,40 +700,49 @@ sockets.updateSimRecord = async function (device_id, response) {
     console.log('entries is: ', response.entries)
 
     let arr = JSON.parse(response.entries);
-    // console.log('arr is: ', arr);
-
-    if (response.action == 'sim_delete') {
-        arr.map(async function (iccID, index) {
-            let dQry = `DELETE FROM sims WHERE device_id = '${device_id}' AND iccid = '${iccID}'`;
-            await sql.query(dQry);
-        })
+    console.log('parsed data is: ', arr);
+    if (response.action == "sim_unregister") {
+        console.log('you are at unReg Section');
+        await sql.query(`UPDATE sims SET unrGuest=${arr.guest}, unrEncrypt=${arr.encrypt} WHERE device_id='${device_id}'`, async function (err, reslt) {
+            if (err) console.log(err)
+        });
     } else {
-        arr.map(async function (data, index) {
-            let guest = 0;
-            let encrypt = 0;
-            if (data.guest) guest = 1; else guest = 0;
-            if (data.encrypt) encrypt = 1; else encrypt = 0;
 
-            let sQry = `SELECT * FROM sims WHERE device_id = '${device_id}' AND iccid = '${data.iccid}'`;
-            console.log(sQry);
-            let rslt = await sql.query(sQry);
-            console.log(rslt);
-
-            console.log('guest is: ', guest)
-            console.log('encrypt is: ', encrypt)
-            if (rslt.length < 1) {
-                let IQry = `INSERT IGNORE INTO sims (device_id, iccid, name, sim_id, note, guest, encrypt, status, dataLimit, sync) 
-    VALUES ('${device_id}', '${data.iccid}', '${data.name}', '', '${data.note}', '${guest}', '${encrypt}', '${data.status}', '', '1');`;
-                await sql.query(IQry, async function (err, result) {
-                    if (err) console.log(err);
+        if (arr.length > 0) {
+            if (response.action == 'sim_delete') {
+                arr.map(async function (iccID, index) {
+                    let dQry = `DELETE FROM sims WHERE device_id = '${device_id}' AND iccid = '${iccID}'`;
+                    await sql.query(dQry);
                 })
             } else {
-                let uQry = `UPDATE sims SET name='${data.name}', note='${data.note}', guest='${guest}', encrypt='${encrypt}', status='${data.status}', sync = '1' WHERE device_id = '${device_id}' AND iccid = '${data.iccid}'`;
-                console.log('UPDATE qry is ', uQry);
-                await sql.query(uQry);
-            }
+                arr.map(async function (data, index) {
+                    // let guest = 0;
+                    // let encrypt = 0;
+                    // if (data.guest) guest = 1; else guest = 0;
+                    // if (data.encrypt) encrypt = 1; else encrypt = 0;
 
-        })
+                    let sQry = `SELECT * FROM sims WHERE device_id = '${device_id}' AND iccid = '${data.iccid}'`;
+                    // console.log(sQry);
+                    let rslt = await sql.query(sQry);
+                    // console.log(rslt);
+
+                    // console.log('guest is: ', guest)
+                    // console.log('encrypt is: ', encrypt)
+                    if (rslt.length < 1) {
+                        let IQry = `INSERT IGNORE INTO sims (device_id, iccid, name, sim_id, note, guest, encrypt, status, dataLimit, sync) 
+    VALUES ('${device_id}', '${data.iccid}', '${data.name}', '', '${data.note}', ${data.guest}, ${data.encrypt}, '${data.status}', '', '1');`;
+                        await sql.query(IQry, async function (err, result) {
+                            if (err) console.log(err);
+                        })
+                    } else {
+                        let uQry = `UPDATE sims SET name='${data.name}', note='${data.note}', guest=${data.guest}, encrypt=${data.encrypt}, status='${data.status}', sync = '1' WHERE device_id = '${device_id}' AND iccid = '${data.iccid}'`;
+                        // console.log('UPDATE qry is ', uQry);
+                        await sql.query(uQry);
+                    }
+
+                })
+            }
+        }
     }
     io.emit(Constants.RECV_SIM_DATA + device_id, {
         status: true
